Add files via upload
A linked list is a data structure consisting of a sequence of nodes, where each node contains two components: the data and a reference (or pointer) to the next node in the sequence. Unlike arrays or lists in Python, linked lists do not store data in contiguous memory locations, making them more dynamic and efficient in terms of memory usage for certain operations.