Add files via upload
an adjacency matrix is a way of representing a graph using a 2D array or matrix. This structure is commonly used to represent relationships between vertices in a graph, where the rows and columns correspond to the vertices, and the values within the matrix indicate the presence or absence of edges between them.